refactor(ingest): Move sqlalchemy import out of sql_types.py
